build: carve llm4s-memory and llm4s-memory-postgres out of core (#1129)
Slice 2 of the modularisation programme (#1126). `org.llm4s.agent.memory` becomes
`llm4s-memory`, except `PostgresMemoryStore`, which becomes `llm4s-memory-postgres`.

Nothing outside the package referenced it, so it moves whole with no facade left behind:
package names are unchanged and this slice has no source breaks. `agent/memory` also
turned out to read no `reference.conf` keys and to be named by no `Llm4sConfig` method,
so there is nothing to migrate on either front.

The stores split across two artifacts because `PostgresMemoryStore` was the package's
only heavyweight - shipping it with `InMemoryStore` would mean every user of agent memory
inherits HikariCP and a JDBC driver. `llm4s-memory` carries sqlite-jdbc for the two
file-backed stores and nothing else.

Core sheds HikariCP and Postgres, and sqlite-jdbc with them. All three were declared in
`commonSettings`, which put a driver and a pool on every module's classpath including
those with no database code; they now sit on the projects that open a connection.

`org.llm4s.vectorstore.PostgresVectorHelpers` stays in `llm4s-core` rather than moving to
`llm4s-rag` with the rest of its package, which retires the temporary duplicate slice 1
left in `org.llm4s.agent.memory`. It is a pure pgvector text codec naming no JDBC type,
and its two consumers - `PgVectorStore` in `llm4s-rag` and `PostgresMemoryStore` in
`llm4s-memory-postgres` - sit in modules that must not depend on each other; having
memory-postgres reach for rag to get twenty lines would drag Tika, POI, PDFBox, jsoup and
the AWS SDK onto the classpath of anyone storing agent memory in Postgres, which is the
opposite of what this carve is for.

Coverage floors are the measured value rounded down to the nearest 5: memory 80 (81.10%),
memory-postgres 60 (60.33%); core measures 73.85%, down from 74.19%, and holds at 70. The
CI coverage job runs and uploads both new modules, each under its own codecov flag.
`modules/it` picks up both in `dependsOn`, so `PostgresMemoryStoreSpec` keeps running on
every PR against the pgvector service container.

Relates to #1126
Closes #1129
