+++ .github/workflows/ci.yml
@@ -176,6 +176,8 @@ jobs:
           core/test core/coverageReport
           rag/test rag/coverageReport
           knowledgegraph/test knowledgegraph/coverageReport
+          memory/test memory/coverageReport
+          memoryPostgres/test memoryPostgres/coverageReport
           coverageAggregate
 
       # One upload per measured module, each carrying that module's flag - Codecov attributes
@@ -207,6 +209,26 @@ jobs:
           token: ${{ secrets.CODECOV_TOKEN }}
           files: modules/knowledgegraph/target/scala-3.7.1/scoverage-report/scoverage.xml
           flags: knowledgegraph
+          slug: llm4s/llm4s
+          fail_ci_if_error: false
+          verbose: true
+
+      - name: Upload memory coverage to Codecov
+        uses: codecov/codecov-action@v7
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: modules/memory/target/scala-3.7.1/scoverage-report/scoverage.xml
+          flags: memory
+          slug: llm4s/llm4s
+          fail_ci_if_error: false
+          verbose: true
+
+      - name: Upload memory-postgres coverage to Codecov
+        uses: codecov/codecov-action@v7
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: modules/memory-postgres/target/scala-3.7.1/scoverage-report/scoverage.xml
+          flags: memory-postgres
           slug: llm4s/llm4s
           fail_ci_if_error: false
           verbose: true
