+++ CLAUDE.md
@@ -27,7 +27,7 @@ Slice order — each is an issue with its own scope and gotchas:
 |---|---|---|
 | 0 ✅ | [#1127](https://github.com/llm4s/llm4s/issues/1127) | build + tracker prerequisites |
 | 1 ✅ | [#1128](https://github.com/llm4s/llm4s/issues/1128) | `llm4s-rag`, `llm4s-knowledgegraph` |
-| 2 | [#1129](https://github.com/llm4s/llm4s/issues/1129) | `llm4s-memory` |
+| 2 ✅ | [#1129](https://github.com/llm4s/llm4s/issues/1129) | `llm4s-memory`, `llm4s-memory-postgres` |
 | 3 | [#1130](https://github.com/llm4s/llm4s/issues/1130) | `llm4s-mcp`, `llm4s-image`, `llm4s-speech` |
 | 4 | [#1131](https://github.com/llm4s/llm4s/issues/1131) | provider registration SPI |
 | 5 | [#1132](https://github.com/llm4s/llm4s/issues/1132) | provider modules |
@@ -58,6 +58,8 @@ llm4s/
 │   ├── core/                  # Core library (published)
 │   ├── rag/                   # RAG, vector stores, chunking, reranking, extraction (published)
 │   ├── knowledgegraph/        # Knowledge graph model, storage, query (published)
+│   ├── memory/                # Agent memory: managers, in-memory + SQLite stores (published)
+│   ├── memory-postgres/       # Agent memory: Postgres/pgvector store (published)
 │   ├── samples/               # Usage examples
 │   ├── workspace/             # Containerized execution
 │   ├── config-policy/         # Config policy checks + CLI
@@ -70,15 +72,22 @@ llm4s/
 └── build.sbt
 ```
 
-Slices 0 and 1 have landed: `modules/rag` and `modules/knowledgegraph` are carved, so
-`modules/core` no longer holds `rag`, `vectorstore`, `chunking`, `reranker`, `eval`,
-`knowledgegraph` or any Tika/POI/PDFBox/jsoup/AWS dependency.
+Slices 0, 1 and 2 have landed: `modules/rag`, `modules/knowledgegraph`, `modules/memory` and
+`modules/memory-postgres` are carved, so `modules/core` no longer holds `rag`, `vectorstore`
+(bar `PostgresVectorHelpers`, see below), `chunking`, `reranker`, `eval`, `knowledgegraph` or
+`agent/memory`, nor any Tika/POI/PDFBox/jsoup/AWS, HikariCP, Postgres or SQLite dependency.
+Those three JDBC dependencies also left `commonSettings`, which used to put them on every
+module's classpath - declare them per-module if you add database code.
+
+`org.llm4s.vectorstore.PostgresVectorHelpers` is the one file in that package still in core:
+it is a pure pgvector text codec shared by `llm4s-rag` and `llm4s-memory-postgres`, which must
+not depend on each other.
 
 **Key paths in `modules/core/src/main/scala/org/llm4s/`:**
 - `types/` - Result type, newtypes
 - `config/` - Llm4sConfig + typed loaders
 - `llmconnect/` - LLM client and providers
-- `agent/` - Agent framework, guardrails, memory, handoffs
+- `agent/` - Agent framework, guardrails, handoffs (memory lives in `modules/memory`)
 - `toolapi/` - Tool calling, built-in tools
 - `trace/` - Observability
