+++ docs/guide/agents/memory.md
@@ -32,6 +32,23 @@ The LLM4S Memory System provides:
 
 ---
 
+## Dependency
+
+{: .note }
+> Not yet published. `llm4s-memory` exists in the build as of
+> [#1129](https://github.com/llm4s/llm4s/issues/1129) but ships in the next release;
+> in `0.4.1` and earlier this code is inside `llm4s-core`.
+
+```scala
+libraryDependencies += "org.llm4s" %% "llm4s-memory" % llm4sVersion
+```
+
+The Postgres backend is a separate artifact - see
+[Postgres Memory Store](../../reference/postgres-memory-store.md). Package names are unchanged
+either way, so `org.llm4s.agent.memory.*` imports keep working.
+
+---
+
 ## Quick Start
 
 ```scala
