+++ modules/memory/src/main/scala/org/llm4s/agent/memory/EmbeddingMemoryStore.scala
