+++ modules/memory/src/main/scala/org/llm4s/agent/memory/SimpleMemoryManager.scala
