+++ modules/memory/src/main/scala/org/llm4s/agent/memory/VectorMemoryStore.scala
