+++ modules/memory/src/test/scala/org/llm4s/agent/memory/MemoryManagerConfigBinaryCompatSpec.scala
